135: Separate PitRunnerMain
Task-Url: http://github.com/pitest/pitclipse/issues/135
