+++ bundles/org.pitest.pitclipse.runner/src/org/pitest/pitclipse/runner/PitRunnerMain.java
@@ -1,0 +1,53 @@
+/*******************************************************************************
+ * Copyright 2012-2019 Phil Glover and contributors
+ *  
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License.  You may obtain a copy
+ * of the License at
+ *  
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ ******************************************************************************/
+
+package org.pitest.pitclipse.runner;
+
+import static com.google.common.base.Preconditions.checkArgument;
+import static java.lang.Integer.parseInt;
+
+import java.io.IOException;
+import java.util.Optional;
+
+import org.pitest.pitclipse.runner.client.PitClient;
+
+/**
+ * Executes PIT providing a Java main, used by the Pitclipse launcher.
+ */
+public class PitRunnerMain {
+
+    public static void main(String[] args) {
+        validateArgs(args);
+        int port = parseInt(args[0]);
+        
+        try (PitClient client = new PitClient(port)) {
+            client.connect();
+            Optional<PitRequest> request = client.readRequest();
+            Optional<PitResults> results = request.map(PitRunner.executePit());
+
+            results.ifPresent(client::sendResults);
+        } catch (IOException e) {
+            // An error occurred while closing the client
+            e.printStackTrace();
+        }
+    }
+
+
+    private static void validateArgs(String[] args) {
+        checkArgument(args.length == 1);
+    }
+
+}
