+++ pom.xml
@@ -57,7 +57,8 @@
 				- untestable activators -->
 		<sonar.exclusions>
 			**/*Exception.java,
-			**/*Activator.java
+			**/*Activator.java,
+			**/*RunnerMain.java,
 		</sonar.exclusions>
 		<sonar.coverage.jacoco.xmlReportPaths>
 			${basedir}/../../tests/org.pitest.pitclipse.tests.coverage.report/target/site/jacoco-aggregate/jacoco.xml
